chore(mobile,shared): Move urql migration-related hooks to shared target
