+++ apps/mobile/src/urql-shared/hooks/urql-migration/useUrqlQueryAction.js
@@ -1,7 +1,7 @@
 import { isString, isObject, isFunction } from 'lodash/fp'
 import { useState, useEffect, useCallback } from 'react'
 import useDeepCompareMemoize from './useDeepCompareMemoize'
-import { fetchGraphqlAction } from 'hooks/useGraphqlAction'
+import { fetchGraphqlAction } from './useGraphqlAction'
 import { useClient } from 'urql'
 import { useDispatch } from 'react-redux'
